This function edits the getPCA function from qsea to allow for batch effect correction

This function edits the getPCA function from qsea to allow for batch effect correction

getPCAwithBatch(
  qs,
  chr = qsea::getChrNames(qs),
  minRowSum = 20,
  minEnrichment = 3,
  keep,
  normMethod = "beta",
  topVar = 1000,
  samples = qsea::getSampleNames(qs),
  batchVariable = NULL
)

Arguments

qs

A qseaSet object

chr

Which chromosomes to use

minRowSum

Minimum number of reads in a window across all the samples to keep it

minEnrichment

Minimal number of expected reads for a fully methylated window (for transformation to absolute methylation level)

keep

a vector of row indices to keep (optional)

normMethod

What normalisation method to use

topVar

Number of most variable windows to keep

samples

Which samples to use

batchVariable

A variable in the sampleTable to batch correct for, using limma::removeBatchEffect