Ultra-stable methylated regions (GRCh38)

Genomic loci that are consistently methylated (“ultra-stable”) across many human tissues and diseases, mapped to GRCh38. Regions correspond to Illumina array probe loci (cg identifiers) derived as in Edgar et al. (2014), and are intended for QC and summary metrics (e.g., addHyperStableFraction()).

hg38UltraStableProbes

Format

A GRanges object with 974 ranges and 1 metadata column:

Probe_ID

Character cg identifier for the probe/locus (e.g., "cg11733071").

Source

Processed from public array resources following Edgar et al. (2014).

Details

Seqnames: numeric chromosomes "1"–"22" (no "chr" prefix; UCSC style can be set if needed).

Strand: "*" for all ranges.

Widths: 3 bp windows (probe CpG ±1 bp).

Genome metadata: currently unset; set to "hg38" if you need an explicit tag.

When combining with other GRanges, harmonise styles/metadata as needed, e.g. use GenomeInfoDb::seqlevelsStyle() to convert to "UCSC" ("chr1", …) and GenomeInfoDb::genome() to set "hg38".

References

Edgar R, Tan PPC, Portales-Casamar E, Pavlidis P (2014). Meta-analysis of human methylomes reveals stably methylated sequences surrounding CpG islands associated with high gene expression. https://pubmed.ncbi.nlm.nih.gov/25493099/

See also

Examples

data(hg38UltraStableProbes)

# Basic overview
hg38UltraStableProbes
#> GRanges object with 974 ranges and 1 metadata column:
#>         seqnames            ranges strand |    Probe_ID
#>            <Rle>         <IRanges>  <Rle> | <character>
#>     [1]        1     951162-951164      * |  cg11733071
#>     [2]        1   1228290-1228292      * |  cg27310827
#>     [3]        1   1547825-1547827      * |  cg24470133
#>     [4]        1   1547876-1547878      * |  cg00355315
#>     [5]        1   2168995-2168997      * |  cg10313879
#>     ...      ...               ...    ... .         ...
#>   [970]       22 46827496-46827498      * |  cg15317817
#>   [971]       22 47024748-47024750      * |  cg18589846
#>   [972]       22 47060889-47060891      * |  cg01578521
#>   [973]       22 49821037-49821039      * |  cg13273562
#>   [974]       22 49888698-49888700      * |  cg22864414
#>   -------
#>   seqinfo: 22 sequences from an unspecified genome; no seqlengths
length(hg38UltraStableProbes)             # 974
#> [1] 974
all(GenomicRanges::width(hg38UltraStableProbes) == 3)  # TRUE (3-bp windows)
#> [1] TRUE

# The object uses numeric chromosomes and an unset genome tag
GenomeInfoDb::seqlevels(hg38UltraStableProbes)[1:5]
#> [1] "1"  "10" "11" "12" "13"
GenomeInfoDb::genome(hg38UltraStableProbes)  # empty
#>  1 10 11 12 13 14 15 16 17 18 19  2 20 21 22  3  4  5  6  7  8  9 
#> NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA 

# Harmonise style/metadata if needed
gr <- hg38UltraStableProbes
GenomeInfoDb::seqlevelsStyle(gr) <- "UCSC"  # or "NCBI" / "Ensembl" for 1,2... instead of chr1, chr2...
GenomeInfoDb::genome(gr) <- "hg38"  # or "GRCh38"

# Count overlaps with a toy region
toy <- GenomicRanges::GRanges("1", IRanges::IRanges(951160, 951170))
GenomicRanges::countOverlaps(toy, hg38UltraStableProbes)
#> [1] 1

# With UCSC-style example
toy_ucsc <- GenomicRanges::GRanges("chr1", IRanges::IRanges(951160, 951170))
GenomicRanges::countOverlaps(toy_ucsc, gr)   # after style harmonisation above
#> [1] 1