This function takes a qseaSet and a GRanges object, and plots the expression across the gene as a heatmap

This function takes a qseaSet and a GRanges object, and plots the expression across the gene as a heatmap

plotGRangesHeatmap(
  qseaSet,
  signatureGR,
  normMethod = "beta",
  annotationCol = NULL,
  annotationColors = NA,
  clusterRows = FALSE,
  clusterCols = TRUE,
  minEnrichment = 3,
  maxScale = 5,
  clusterNum = 2,
  description = "",
  clip = 1000000000,
  minDensity = 0,
  clusterMethod = "ward.D2",
  ...
)

Arguments

qseaSet

The qseaSet object.

signatureGR

A genomic ranges to plot.

normMethod

Whether to plot nrpm values or beta values.

annotationCol

A data frame with annotations for the samples.

annotationColors

A list specifying some or all of the colours to use for the annotations.

clusterRows

Whether to cluster the rows or not.

clusterCols

Whether to cluster the columns or not.

minEnrichment

Minimum enrichment factor for beta values, will give NAs below this.

maxScale

The maximum of the scale, not used when plotting beta values.

clusterNum

A number of clusters to break the column dendrogram into.

description

A string to include as part of the title.

clip

A level to clip the data at, anything higher than this is replaced with the clip value.

minDensity

A minimum CpG density level to filter out windows with values lower than.

clusterMethod

What method to use to cluster the dendrograms.

...

Other arguments to pass to pheatmap.

Value

A qseaSet object with the sampleTable enhanced with the information on number of reads etc