This function takes a qseaSet and plots a PCA

This function takes a qseaSet and plots a PCA

plotQseaPCA(
  qseaSet,
  signatureGR = NULL,
  plotComponents = c(1, 2),
  plotColour = "type",
  plotShape = "experiment",
  batchVariable = NULL,
  showNames = FALSE,
  normMethod = "beta",
  minRowSum = 20,
  minEnrichment = 3,
  topVar = NULL,
  plotTitle = "",
  returnDataOnly = FALSE
)

Arguments

qseaSet

The qseaSet object.

signatureGR

A GRanges (or data frame coercible to one) with a subset of windows to calculate the PCA on.

plotComponents

Which of the PCA components to plot default c(1,2)

plotColour

Column of the sampleTable to use to set the colour of the points

plotShape

Column of the sampleTable to use to set the shape of the points

batchVariable

A variable to use for batch correction using limma. Can be overaggressive and remove all variation. Be careful with this!

showNames

Whether to show the names of the points

normMethod

Normalisation method to use, nrpm or beta.

minRowSum

A minimum number of reads to be included in the PCA

minEnrichment

Minimal number of expected reads for a fully methylated window (for transformation to absolute methylation level)

topVar

Take only the n most variable windows in the PCA. If NULL, all valid windows.

plotTitle

Title for the plot

returnDataOnly

Return just the data frame behind the PCA

Value

A plot of the PCA