Calculate and add CNV based on HMMcopy, on individual samples

This calculates and adds CNV to a qseaSet based on just using HMMcopy on each sample, with default parameters. Only works for hg38/GRCh38!

addHMMcopyCNV(
  qs,
  inputColumn = "input_file",
  windowSize = 1000000,
  fragmentLength = NULL,
  plotDir = NULL,
  parallel = TRUE,
  maxInsertSize = 1000,
  minInsertSize = 50,
  minReferenceLength = 30,
  minMapQual = 30,
  properPairsOnly = FALSE
)

Arguments

qs

The qseaSet object

inputColumn

Which column of the sample table contains the link to the bam file

windowSize

What window size to use (only 50000, 500000 or 1000000 work at the moment)

fragmentLength

Fragment length to extend reads to if not paired

plotDir

A directory to export individual HMMcopy plots to.

parallel

Whether to use parallelisation

maxInsertSize

For paired reads, only keep them if they are below a maximum length. Can be used for cfDNA size selection. Applies to Input samples as well as MeCap.

minInsertSize

For paired reads, only keep them if they are above a minimum length. Can be used for cfDNA size selection. Applies to Input samples as well as MeCap.

minReferenceLength

A minimum distance on the genome to keep the read. bwa by default gives 19bp as minimum for a read, which is quite short.

minMapQual

Minimum mapping quality to include a read (on either end if proper pair)

properPairsOnly

Whether to only keep properly paired reads, or to keep high-quality (MAPQ 30+) unpaired R1s as well. Set to TRUE for size selection.

Value

A qseaGLM object