Combine two qseaSets

This function takes a list of qseaSets (or strings to read rds files from) and merges them together

combineQsetsList(
  qseaSets,
  firstQset = NULL,
  dropDuplicates = TRUE,
  checkParams = TRUE,
  regionsToKeep = NULL
)

Arguments

qseaSets

A list of qseaSet objects or strings to rds files with them in

firstQset

An first qseaSet object to combine the rest into (optional)

dropDuplicates

Whether to drop samples with the same name. Else renames them by adding "_Dup" to the end of the name.

checkParams

Whether to verify that all the parameters are identical in both objects

regionsToKeep

A GRanges object (or table coercible to one) to use to subset the samples. Helps keep RAM use down.

Value

A qseaSet object, containing all the samples from both qseaSet objects.