Make the initial qseaSet object from a set of samples

This function makes the initial qseaSet from a sampleTable that includes the file paths for the bam files and their metadata

makeQset(
  sampleTable,
  BSgenome = NULL,
  chrSelect = 1:22,
  windowSize = 300,
  CNVwindowSize = 1e+06,
  fragmentType = NULL,
  fragmentLength = NULL,
  fragmentSD = NULL,
  CNVmethod = "HMMdefault",
  coverageMethod = "PairedAndR1s",
  minMapQual = 10,
  minInsertSize = 70,
  maxInsertSize = 1000,
  minReferenceLength = 30,
  badRegions = NULL,
  properPairsOnly = FALSE
)

Arguments

sampleTable

A data frame with the sample information to be passed to qsea.

BSgenome

A BSgenome string. See BSgenome::available.genomes() for options.

chrSelect

Which chromosomes to use (default 1:22).

windowSize

What window size (in bp) to use in the genome (default 300).

CNVwindowSize

What window size (in bp) to use in the calculation of CNV (default 1000000).

fragmentType

What type of procedure generated the library, Sheared or cfDNA (used to set the average fragment length/SD to a default for CBC)

fragmentLength

Average DNA fragment length. Can be set by fragmentType.

fragmentSD

Standard deviation of the DNA fragment lengths. Can be set by fragmentType.

CNVmethod

Which method to use for calculation of CNV. Options include "HMMdefault" (hmmcopy with default parameters) and "None"

coverageMethod

Whether to use custom methtools method for reading coverage in (set to PairedAndR1s), rather than qsea's (qseaPaired).

minMapQual

Minimum MAPQ score for a read to be kept. For PairedAndR1s, will keep if either R1 or R2 meet this cutoff in a properly paired read (if MQ tags are set in the bam file with samtools fixmate).

minInsertSize

For paired reads, only keep them if they are above a minimum length. Can be used for cfDNA size selection. Applies to Input samples as well as MeCap.

maxInsertSize

For paired reads, only keep them if they are below a maximum length. Can be used for cfDNA size selection. Applies to Input samples as well as MeCap.

minReferenceLength

A minimum distance on the genome to keep the read. bwa by default gives 19bp as minimum for a read, which is quite short.

badRegions

A GRanges object containing regions to filter out from the result.

properPairsOnly

Whether to only keep properly paired reads, or to keep high-quality (MAPQ 30+) unpaired R1s as well. Set to TRUE for size selection.

Value

A qseaSet object, containing all the information required.